Many important updates (#2)
* Added load_pnl functionality

* Type hint fix for num_portfolios

* Updated poetry files

* Updated requirements.yml

* Updated examples to include efficient frontiers

* Added license text to all source files

* Added doc string to load_pnl

* Updated examples

+++ fortitudo/tech/optimization.py
@@ -212,7 +212,7 @@ class MeanCVaR:
         else:
             raise ValueError('Constraints are infeasible or _max_expected_return is unbounded.')
 
-    def efficient_frontier(self, num_portfolios: float = None) -> np.ndarray:
+    def efficient_frontier(self, num_portfolios: int = None) -> np.ndarray:
         """Method for computing the efficient frontier.
 
         Args:
